refactor: ShowMoreButton into IntegrationPlatformsSection folder
